# apps/server-api/docs/swagger.yaml
basePath: /
definitions:
...
  http.AdminOverviewResponse:
    properties:
      activeStudentCount:
        example: 172
        type: integer
      activeTeacherCount:
        example: 7
        type: integer
      adminCount:
        example: 1
        type: integer
      disabledStudentCount:
        example: 8
        type: integer
      disabledTeacherCount:
        example: 1
        type: integer
      studentCount:
        example: 180
        type: integer
      teacherCount:
        example: 8
        type: integer
    type: object
  http.AdminStudentItemResponse:
    properties:
      createdAt:
        example: "2026-06-13T12:00:00Z"
        type: string
      displayName:
        example: 小明
        type: string
      id:
        example: 1
        type: integer
      status:
        example: active
        type: string
      teacherId:
        example: 2
        type: integer
      teacherUsername:
        example: teacher01
        type: string
      username:
        example: student01
        type: string
    type: object
  http.AdminStudentPasswordResetRequestDoc:
    properties:
      newPassword:
        example: updated123
        type: string
    required:
    - newPassword
    type: object
  http.AdminStudentsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.AdminStudentItemResponse'
        type: array
    type: object
  http.AdminTeacherCreateRequestDoc:
    properties:
      initialPassword:
        example: secret123
        type: string
      username:
        example: teacher01
        type: string
    required:
    - initialPassword
    - username
    type: object
  http.AdminTeacherItemResponse:
    properties:
      createdAt:
        example: "2026-06-13T12:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      role:
        example: teacher
        type: string
      status:
        example: active
        type: string
      username:
        example: teacher01
        type: string
    type: object
  http.AdminTeacherPasswordResetRequestDoc:
    properties:
      newPassword:
        example: updated123
        type: string
    required:
    - newPassword
    type: object
  http.AdminTeachersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.AdminTeacherItemResponse'
        type: array
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
...
info:
  contact: {}
  description: OpenAPI contract for the Scratch AI teaching server.
  title: Scratch AI Server API
  version: "1.0"
paths:
  /api/admin/overview:
    get:
      description: Read aggregated teacher/student account counts for the current
        administrator.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminOverviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin overview
      tags:
      - admin
  /api/admin/students:
    get:
      description: List all managed student accounts for the current administrator.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminStudentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List students
      tags:
      - admin
  /api/admin/students/{id}/disable:
    post:
      description: Disable one managed student account.
      parameters:
      - description: Student ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminStudentItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable student
      tags:
      - admin
  /api/admin/students/{id}/enable:
    post:
      description: Enable one managed student account.
      parameters:
      - description: Student ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminStudentItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enable student
      tags:
      - admin
  /api/admin/students/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: Replace one managed student account password.
      parameters:
      - description: Student ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: New password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.AdminStudentPasswordResetRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminStudentItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reset student password
      tags:
      - admin
  /api/admin/teachers:
    get:
      description: List all teacher/admin accounts for the current administrator.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminTeachersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List teachers
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a managed teacher account with an initial password.
      parameters:
      - description: Teacher create payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.AdminTeacherCreateRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.AdminTeacherItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create teacher
      tags:
      - admin
  /api/admin/teachers/{id}/disable:
    post:
      description: Disable one managed teacher account.
      parameters:
      - description: Teacher ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminTeacherItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable teacher
      tags:
      - admin
  /api/admin/teachers/{id}/enable:
    post:
      description: Enable one managed teacher account.
      parameters:
      - description: Teacher ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminTeacherItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enable teacher
      tags:
      - admin
  /api/admin/teachers/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: Replace one managed teacher account password.
      parameters:
      - description: Teacher ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: New password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.AdminTeacherPasswordResetRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminTeacherItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reset teacher password
      tags:
      - admin
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"